Fix for issue #93
While the nodes start up, it can happen that a point cloud arrives that does not find a matching pose. This leads to an Error and confusion of the user. The system starts to work normally once a matching point cloud and pose is available.
An initializing functionality that waits for a matching pair of point cloud and pose is implemented with this commit.\
The user is indicated with a Warn message that initialization is not finished.
The original Error message has been extended for better readability.
Note: robotPoseCache_ does not have a function to check if it is empty.